requirements.txt
flask==3.1.1
python-dotenv==1.1.0
google-cloud-texttospeech==2.27.0
google-generativeai==0.8.5 